+++ src/components/Card/index.tsx
@@ -2,6 +2,7 @@ import React, { memo } from 'react';
 import Style from './style';
 
 type CardProps = {
+  ability1Title?: string;
   active: boolean;
   id: string;
   handleCardMouseDown: (id: string) => void;
@@ -10,6 +11,7 @@ type CardProps = {
 };
 
 const Card = ({
+  ability1Title,
   active,
   id,
   handleCardMouseDown,
@@ -21,7 +23,8 @@ const Card = ({
     onClick={(event: React.MouseEvent<HTMLInputElement>): void => event.stopPropagation()}
     onMouseDown={(): void => handleCardMouseDown(id)}>
     <p className='title'>{title}</p>
-    <p className={'subtitle'}>{subtitle}</p>
+    <p className={'subtext subtitle'}>{subtitle}</p>
+    {ability1Title && <p className='subtext ability1Title'>{ability1Title}</p>}
   </Style>;
 
 export default memo(Card);